Revert "Shuffling around types"
This reverts commit 0ea60c7742fced60f4d7adc4ecb80184bc627150.
