+++ packages/mux-uploader/src/index.ts
@@ -1,6 +1,5 @@
 export * as constants from './constants';
-import { MuxUploaderElementEventMap } from './types';
-import MuxUploaderElement from './mux-uploader';
+import MuxUploaderElement, { MuxUploaderElementEventMap } from './mux-uploader';
 import MuxUploaderProgressElement from './mux-uploader-progress';
 import MuxUploaderDropElement from './mux-uploader-drop';
 import MuxUploaderFileSelectElement from './mux-uploader-file-select';
