+++ packages/mux-uploader/src/mux-uploader.ts
@@ -1,5 +1,4 @@
 import { globalThis, document } from 'shared-polyfills';
-import { MuxUploaderElement as IMuxUploaderElement } from './types';
 import * as UpChunk from '@mux/upchunk';
 
 const styles = `
@@ -110,7 +109,56 @@ const ButtonPressedKeys = ['Enter', ' '];
 type Endpoint = UpChunk.UpChunk['endpoint'] | undefined | null;
 type DynamicChunkSize = UpChunk.UpChunk['dynamicChunkSize'] | undefined;
 
-class MuxUploaderElement extends globalThis.HTMLElement implements IMuxUploaderElement {
+type ErrorDetail = {
+  message: string;
+  chunkNumber?: number;
+  attempts?: number;
+};
+
+// NOTE: Progress event is already determined on HTMLElement but have inconsistent types. Should consider renaming events (CJP)
+export interface MuxUploaderElementEventMap extends Omit<HTMLElementEventMap, 'progress'> {
+  uploadstart: CustomEvent<{ file: File; chunkSize: number }>;
+  chunkattempt: CustomEvent<{
+    chunkNumber: number;
+    chunkSize: number;
+  }>;
+  chunksuccess: CustomEvent<{
+    chunk: number;
+    chunkSize: number;
+    attempts: number;
+    timeInterval: number;
+    // Note: This should be more explicitly typed in Upchunk. (TD).
+    response: any;
+  }>;
+  uploaderror: CustomEvent<ErrorDetail>;
+  progress: CustomEvent<number>;
+  success: CustomEvent<undefined | null>;
+}
+
+interface MuxUploaderElement extends HTMLElement {
+  addEventListener<K extends keyof MuxUploaderElementEventMap>(
+    type: K,
+    listener: (this: HTMLMediaElement, ev: MuxUploaderElementEventMap[K]) => any,
+    options?: boolean | AddEventListenerOptions
+  ): void;
+  addEventListener(
+    type: string,
+    listener: EventListenerOrEventListenerObject,
+    options?: boolean | AddEventListenerOptions
+  ): void;
+  removeEventListener<K extends keyof MuxUploaderElementEventMap>(
+    type: K,
+    listener: (this: HTMLMediaElement, ev: MuxUploaderElementEventMap[K]) => any,
+    options?: boolean | EventListenerOptions
+  ): void;
+  removeEventListener(
+    type: string,
+    listener: EventListenerOrEventListenerObject,
+    options?: boolean | EventListenerOptions
+  ): void;
+}
+
+class MuxUploaderElement extends globalThis.HTMLElement implements MuxUploaderElement {
   protected _formatProgress: ((percent: number) => string) | null | undefined;
   protected _endpoint: Endpoint;
   uploadPercentage: HTMLElement | null | undefined;
