--- packages/mux-uploader/src/types.ts
@@ -1,48 +1,0 @@
-type ErrorDetail = {
-  message: string;
-  chunkNumber?: number;
-  attempts?: number;
-};
-
-// NOTE: Progress event is already determined on HTMLElement but have inconsistent types. Should consider renaming events (CJP)
-export interface MuxUploaderElementEventMap extends Omit<HTMLElementEventMap, 'progress'> {
-  uploadstart: CustomEvent<{ file: File; chunkSize: number }>;
-  chunkattempt: CustomEvent<{
-    chunkNumber: number;
-    chunkSize: number;
-  }>;
-  chunksuccess: CustomEvent<{
-    chunk: number;
-    chunkSize: number;
-    attempts: number;
-    timeInterval: number;
-    // Note: This should be more explicitly typed in Upchunk. (TD).
-    response: any;
-  }>;
-  uploaderror: CustomEvent<ErrorDetail>;
-  progress: CustomEvent<number>;
-  success: CustomEvent<undefined | null>;
-}
-
-export interface MuxUploaderElement extends HTMLElement {
-  addEventListener<K extends keyof MuxUploaderElementEventMap>(
-    type: K,
-    listener: (this: HTMLMediaElement, ev: MuxUploaderElementEventMap[K]) => any,
-    options?: boolean | AddEventListenerOptions
-  ): void;
-  addEventListener(
-    type: string,
-    listener: EventListenerOrEventListenerObject,
-    options?: boolean | AddEventListenerOptions
-  ): void;
-  removeEventListener<K extends keyof MuxUploaderElementEventMap>(
-    type: K,
-    listener: (this: HTMLMediaElement, ev: MuxUploaderElementEventMap[K]) => any,
-    options?: boolean | EventListenerOptions
-  ): void;
-  removeEventListener(
-    type: string,
-    listener: EventListenerOrEventListenerObject,
-    options?: boolean | EventListenerOptions
-  ): void;
-}
